+++ packages/popmotion-pose/src/index.ts
@@ -1,3 +1,5 @@
 import domPose from './dom/pose';
+import { setReducedMotion } from './inc/accessibility';
 
 export default domPose;
+export { setReducedMotion };