+++ packages/react-pose/src/index.ts
@@ -1,5 +1,6 @@
 import posed from './posed';
 import { PoseGroup } from './components/PoseGroup';
+import { setReducedMotion } from 'popmotion-pose';
 
 export default posed;
-export { PoseGroup };
+export { PoseGroup, setReducedMotion };